IA map update for Transform rule:
General change:
Clicked coordinates are now always available if there is a current order set.

Variant/Rule specific:
Updated broken transform script to work again together with order menu toggling.
Behaviour to insert Transform order to a fleet on a territory with several coasts is now the same as for Builds: The coast closest to the clicked coordinates is selected.